refactor(auth): encapsulate the token generation inside a plugin
